pnfs: Imlement layout vnode db layer
This layer should be used for find/store layouts for vnode.

Why it was introduced? Because we need to get layout for vnode for
NFS v3, v4 and v41.

Currently only NFSv41 can path can obtain layout because "instp" is used for
rfs4_findfile.

To get layout in old scheme:
   1. find rfs4_file_t (with using "instp")
   2. use fp->rf_mlo to get/hash layout

Problem is:
          for nfs4 and nfs41 - different rfs4_file_t
          NFSv3 doesn't use/allocate rfs4_file_t

So we need tool to get layout by vnode.

To use layout:

   pnfs_get_mds_layout   -  Get layout by vnode: If layout
                            is in a database, return from hash; Otherwise
                            read layout from layout file;

   pnfs_add_mds_layout   -  Add layout to database and store to disk;

   pnfs_save_mds_layout  -  Store layout to disk;

   pnfs_delete_mds_layout - Delete layout from database and disk.