nfs4[1]: Introduced functions that simplify cs->vp/cs->fh updates.
rfs4_cs_* functions family was introduced in order to simplify
update and invalidate operations on filehandle and vnode (either current
or saved) stored in compound state structure.
New API on compound state structure introduces a criteria of filehandle
correctness. Now filehandle is assumed to be correct if its length is not 0.