pnfs-layout: remove pnfs_save_mds_layout()
For old scheme, when were used two files {metadata and layout}, it was allowed
to save layout later. Now only one file is used and it is strongly required to
save layout during file creating.

Otherwise, we have ds_checkstate process sleeping because this process tried to
write to file that had been delegated.